myproject/requirements.txt
Django==4.2.2
django==4.2.2
django-cors-headers==3.7.0
django-crontab==0.7.1
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
celery==5.3.1
redis==3.5
# rabbitmq==1.15.0